requirements.txt
pyyaml
watchdog
rich
flask